refactor: entity 패키지 분리
