+++ src/main/java/roomescape/domain/ReservationBase.java
@@ -8,6 +8,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.MappedSuperclass;
+import roomescape.entity.Member;
 
 @MappedSuperclass
 public abstract class ReservationBase {
