+++ src/main/java/roomescape/entity/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.entity;
 
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
@@ -7,6 +7,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import roomescape.domain.ThemeName;
 
 @Entity
 public class Theme {
