+++ src/main/java/roomescape/entity/Waiting.java
@@ -1,9 +1,11 @@
-package roomescape.domain;
+package roomescape.entity;
 
 import jakarta.persistence.Entity;
+import roomescape.domain.ReservationBase;
+import roomescape.domain.Schedule;
 
 @Entity
-public class Waiting extends ReservationBase{
+public class Waiting extends ReservationBase {
 
     public Waiting(final Member member, final Schedule schedule) {
         super(member, schedule);
