+++ src/main/java/roomescape/infrastructure/TokenProvider.java
@@ -8,7 +8,7 @@ import java.util.Arrays;
 import javax.crypto.SecretKey;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.domain.Member;
+import roomescape.entity.Member;
 import roomescape.domain.Role;
 import roomescape.handler.exception.CustomException;
 import roomescape.handler.exception.ExceptionCode;
