+++ src/main/java/roomescape/repository/MemberRepository.java
@@ -2,7 +2,7 @@ package roomescape.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.domain.Member;
+import roomescape.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
