+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,7 +1,7 @@
 package roomescape.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.ReservationTime;
 
 public interface ReservationTimeRepository extends JpaRepository<ReservationTime, Long> {
 }
