+++ src/main/java/roomescape/service/WaitingService.java
@@ -4,12 +4,12 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.Member;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.Member;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationTime;
 import roomescape.domain.Schedule;
-import roomescape.domain.Theme;
-import roomescape.domain.Waiting;
+import roomescape.entity.Theme;
+import roomescape.entity.Waiting;
 import roomescape.domain.WaitingWithRank;
 import roomescape.handler.exception.CustomException;
 import roomescape.handler.exception.ExceptionCode;
