+++ src/main/java/roomescape/service/dto/request/ReservationTimeRequest.java
@@ -2,7 +2,7 @@ package roomescape.service.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.ReservationTime;
 
 public record ReservationTimeRequest(
         @NotNull(message = "날짜를 입력해주세요.")
