+++ src/main/java/roomescape/service/dto/response/AvailableTimeResponse.java
@@ -2,7 +2,7 @@ package roomescape.service.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.ReservationTime;
 
 public record AvailableTimeResponse(@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm")
                                     LocalTime time, Long timeId, boolean isAvailable) {
