+++ src/main/java/roomescape/service/dto/response/ReservationTimeResponse.java
@@ -2,7 +2,7 @@ package roomescape.service.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.entity.ReservationTime;
 
 public record ReservationTimeResponse(Long id,
                                       @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm")
